requirements-dev.txt
# Local specific requirements
awscli
fastapi-cli[standard]>=0.0.5
# ----------------------------
fastapi==0.115.0
mangum==0.17.0
aws-cdk-lib==2.159.1
aws-cdk.cx-api==0.28.0
aws-cdk.aws-lambda==0.28.0
aws-cdk.aws-apigateway==0.28.0
constructs>=10.0.0,<11.0.0